@@ -3,19 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "0.21.0",
-    "bootstrap": "^4.5.3",
-    "canvas": "2.6.1",
-    "eslint": "^7.13.0",
-    "@fortawesome/fontawesome-free": "^5.15.1",
-    "gif-encoder-2": "1.0.5",
-    "moment": "^2.29.1",
-    "ol": "^6.4.3",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.0"
+    "axios": "^0.27.2",
+    "bootstrap": "^5.1.3",
+    "canvas": "2.9.1",
+    "eslint": "^8.15.0",
+    "@fortawesome/fontawesome-free": "^6.1.1",
+    "moment": "^2.29.3",
+    "ol": "^6.14.1",
+    "react": "^18.1.0",
+    "react-bootstrap": "^2.4.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -39,6 +38,6 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^6.6.0"
+    "eslint": "^8.15.0"
   }
 }
