@@ -2,6 +2,7 @@ import React from 'react';
 import './App.css';
 
 // react-bootstrap imports
+import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 import NavDropdown from 'react-bootstrap/NavDropdown';
@@ -9,13 +10,12 @@ import NavDropdown from 'react-bootstrap/NavDropdown';
 // react-router-dom imports
 import {
   BrowserRouter as Router,
-  Switch,
+  Routes,
   Route
 } from 'react-router-dom';
 
 // Local imports
 import Home from './pages/home/Home.js';
-import About from './pages/about/About.js';
 import Posts from './pages/posts/Posts.js';
 import Goes16serverless from './pages/goes16serverless/Main.js';
 
@@ -24,39 +24,29 @@ function App() {
     <Router>
       <div>
         <Navbar bg="dark" variant="dark" expand="lg">
-          <Navbar.Brand href="/home">nathancalandra.cloud</Navbar.Brand>
-          <Navbar.Toggle aria-controls="basic-navbar-nav" />
-          <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="mr-auto">
-              <Nav.Link href="/home">Home</Nav.Link>
-              <Nav.Link href="/about">About Me</Nav.Link>
-              <NavDropdown title="Projects" id="basic-nav-dropdown">
-                <NavDropdown.Item href="/goes16-serverless">GOES 16 Serverless</NavDropdown.Item>
-              </NavDropdown>
-            </Nav>
-            <Nav>
-              <Nav.Link href="https://github.com/ncalandra">GitHub</Nav.Link>
-              <Nav.Link href="https://www.linkedin.com/in/nathan-calandra/">LinkedIn</Nav.Link>
-            </Nav>
-          </Navbar.Collapse>
+          <Container>
+            <Navbar.Brand href="/home">nathancalandra.cloud</Navbar.Brand>
+            <Navbar.Toggle aria-controls="basic-navbar-nav" />
+            <Navbar.Collapse id="basic-navbar-nav">
+              <Nav className="mr-auto">
+                <Nav.Link href="/home">Home</Nav.Link>
+                <NavDropdown title="Projects" id="basic-nav-dropdown">
+                  <NavDropdown.Item href="/goes16-serverless">GOES 16 Serverless</NavDropdown.Item>
+                </NavDropdown>
+              </Nav>
+              <Nav>
+                <Nav.Link href="https://github.com/ncalandra">GitHub</Nav.Link>
+                <Nav.Link href="https://www.linkedin.com/in/nathan-calandra/">LinkedIn</Nav.Link>
+              </Nav>
+            </Navbar.Collapse>
+          </Container>
         </Navbar>
-
-        {/* A <Switch> looks through its children <Route>s and
-            renders the first one that matches the current URL. */}
-        <Switch>
-          <Route path="/about">
-            <About />
-          </Route>
-          <Route path="/posts">
-            <Posts />
-          </Route>
-          <Route path="/goes16-serverless">
-            <Goes16serverless />
-          </Route>
-          <Route path="/">
-            <Home />
-          </Route>
-        </Switch>
+        <Routes>
+          <Route path="/posts" element={<Posts />} />
+          <Route path="/goes16-serverless" element={<Goes16serverless />} />
+          <Route path="/" element={<Home />} />
+          <Route path="/home" element={<Home />} />
+        </Routes>
       </div>
     </Router>
   );
